# api/openapi.yaml
openapi: 3.1.0
info:
  title: Waveguide
  version: 1.0.0
  description: |
    Live TV and DVR server for HDHomeRun tuners. This document is the contract for the
    web app and the iOS and tvOS apps; generated clients come from it.

    Every route is served under `/api/v1`. The same routes are also mounted under `/api`
    for the current web app and will be removed once it moves to `/api/v1`.

    Errors use one envelope: `{"code": "tuners_busy", "message": "...", ...details}`.
    Tuner stream URLs and tuner DeviceAuth are never returned.
servers:
  - url: http://localhost:8477/api/v1
...
paths:
...
  /server:
    get:
      tags: [server]
      operationId: getServer
      description: Identity and capabilities. Clients use `id` to recognize a paired server at a new address.
      responses:
        "200":
          description: Server info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerInfo" }
    patch:
      tags: [server]
      operationId: renameServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 63 }
      responses:
        "200":
          description: Updated server info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerInfo" }
        "400": { $ref: "#/components/responses/Error" }
...
  /recordings/{id}/play:
    post:
      tags: [recordings]
      operationId: playRecording
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                pictureMode: { $ref: "#/components/schemas/PictureMode" }
      responses:
        "200":
          description: Playback for a finished or in-progress recording
          content:
            application/json:
              schema:
                type: object
                required: [playlist, position, growing, markers]
                properties:
                  playlist: { type: string }
                  position: { type: number, description: "Resume position in seconds" }
                  growing: { type: boolean, description: "True while the recording is still being written" }
                  markers:
                    type: array
                    items: { $ref: "#/components/schemas/Marker" }
        "404": { $ref: "#/components/responses/Error" }
  /recordings/{id}/progress:
    put:
      tags: [recordings]
      operationId: saveProgress
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [position]
              properties:
                position: { type: number }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  position: { type: number }
...
  /recordings/{id}/file:
    get:
      tags: [recordings]
      operationId: downloadRecording
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The original broadcast
          content:
            video/mp2t:
              schema: { type: string, format: binary }
  /recordings/{id}/markers:
    get:
      tags: [recordings]
      operationId: listMarkers
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Commercial markers
          content:
            application/json:
              schema:
                type: object
                required: [markers]
                properties:
                  markers:
                    type: array
                    items: { $ref: "#/components/schemas/Marker" }
    post:
      tags: [recordings]
      operationId: addMarker
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [start, end]
              properties:
                start: { type: number }
                end: { type: number }
      responses:
        "200":
          description: Marker added
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Marker" }
  /recordings/{id}/detect:
    post:
      tags: [recordings]
      operationId: detectBreaks
      description: Find commercial breaks (comskip when installed, otherwise black frames).
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Markers after detection
          content:
            application/json:
              schema:
                type: object
                required: [markers]
                properties:
                  markers:
                    type: array
                    items: { $ref: "#/components/schemas/Marker" }
  /markers/{id}:
    delete:
      tags: [recordings]
      operationId: deleteMarker
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { $ref: "#/components/responses/Ok" }
...
  /passes/{id}:
    patch:
      tags: [dvr]
      operationId: updatePass
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Pass" }
      responses:
        "200":
          description: Passes after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassList" }
    delete:
      tags: [dvr]
      operationId: deletePass
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Passes after the change
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PassList" }
...
  /virtuals:
    get:
      tags: [virtuals]
      operationId: listVirtuals
      responses:
        "200":
          description: Library channels
          content:
            application/json:
              schema:
                type: object
                required: [virtuals]
                properties:
                  virtuals:
                    type: array
                    items: { $ref: "#/components/schemas/VirtualChannel" }
    post:
      tags: [virtuals]
      operationId: createVirtual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                number: { type: string }
                name: { type: string }
                recordings:
                  type: array
                  items: { type: integer, format: int64 }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualChannel" }
...
  /virtuals/{id}/play:
    post:
      tags: [virtuals]
      operationId: playVirtual
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                index: { type: integer }
                pictureMode: { $ref: "#/components/schemas/PictureMode" }
      responses:
        "200":
          description: Playback for one recording on the channel. Uses no tuner.
          content:
            application/json:
              schema:
                type: object
                properties:
                  usesTuner: { type: boolean }
                  index: { type: integer }
                  count: { type: integer }
                  playlist: { type: string }
                  number: { type: string }
                  name: { type: string }
                  recording: { $ref: "#/components/schemas/Recording" }
                  markers:
                    type: array
                    items: { $ref: "#/components/schemas/Marker" }
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  responses:
    Ok:
      description: Done
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      additionalProperties: true
      properties:
        code:
          type: string
          description: Stable machine code, for example tuners_busy, disk_low, not_found, bad_request.
        message: { type: string, description: Short text safe to show a user. }
        tuners:
          type: array
          items: { $ref: "#/components/schemas/Tuner" }
        freeBytes: { type: integer, format: int64 }
        needBytes: { type: integer, format: int64 }
...
    ServerInfo:
      type: object
      required: [id, name, version, apiVersion, features]
      properties:
        id: { type: string }
        name: { type: string }
        version: { type: string }
        apiVersion: { type: integer }
        encoder: { type: string }
        tunerCount: { type: integer }
        features:
          type: array
          items: { type: string }
...
    PictureMode:
      type: string
      enum: [broadcast, smooth, film]
      description: broadcast rebuilds interlaced video at 60 frames a second; smooth uses motion-compensated deinterlacing; film recovers 24p.
    Tuner:
      type: object
      required: [index, ours]
      properties:
        index: { type: integer }
        guide: { type: string }
        name: { type: string }
        target: { type: string }
        ours: { type: boolean }
        strength: { type: integer }
        quality: { type: integer }
        symbol: { type: integer }
        viewers: { type: integer }
...
    Recording:
      type: object
      required: [id, channelId, guideNumber, title, status, startedAt]
      properties:
        id: { type: integer, format: int64 }
        channelId: { type: integer, format: int64 }
        guideNumber: { type: string }
        title: { type: string }
        subtitle: { type: string }
        description: { type: string }
        category: { type: string }
        programId: { type: string }
        status: { type: string, enum: [recording, complete, failed, stopped, imported] }
        error: { type: string }
        startedAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time }
        bytes: { type: integer, format: int64 }
        position: { type: number }
        durationSec: { type: number }
        watched: { type: integer }
    Marker:
      type: object
      required: [id, start, end]
      properties:
        id: { type: integer, format: int64 }
        recordingId: { type: integer, format: int64 }
        start: { type: number }
        end: { type: number }
    Pass:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        channelId: { type: integer, format: int64 }
        kind: { type: string }
        padBefore: { type: integer, minimum: 0, maximum: 30 }
        padAfter: { type: integer, minimum: 0, maximum: 30 }
        priority: { type: integer }
        episodes: { type: string, enum: [all, new] }
        keepMode: { type: string, enum: [all, latest] }
        keepCount: { type: integer }
        limitCount: { type: integer }
        rerecord: { type: boolean }
        commercials: { type: boolean }
        timeStart: { type: string, description: "HH:MM local" }
        timeEnd: { type: string, description: "HH:MM local" }
        matchKind: { type: string, enum: [title, contains, category] }
    PassList:
      type: object
      required: [passes]
      properties:
        passes:
          type: array
          items: { $ref: "#/components/schemas/Pass" }
...
    VirtualChannel:
      type: object
      required: [id, number, name, recordings]
      properties:
        id: { type: integer, format: int64 }
        number: { type: string }
        name: { type: string }
        orderMode: { type: string }
        ruleTitle: { type: string }
        recordings:
          type: array
          items: { type: integer, format: int64 }
...